Phase 6: real Xcode project builds -- first successful iOS build
Checks in the real Xcode project (iosApp/iosApp/iosApp.xcodeproj and its
Swift sources) now that it actually produces a working iosApp.app for the
iOS Simulator, plus every fix that got it there:

- Kotlin compiler bump 2.2.10 -> 2.3.21 (root build.gradle.kts): current
  published iOS klibs for Ktor 3.5.2 and Compose Multiplatform 1.12.0 are
  built with newer Kotlin than this project started on; a newer compiler
  reading older-ABI klibs is the safe direction.
- androidApp/build.gradle.kts: migrated the old `kotlinOptions { jvmTarget
  = "17" }` DSL to the modern `compilerOptions` block -- the old setter is
  a hard compile error (not just a warning) under Kotlin 2.3.21.
- shared/build.gradle.kts: dropped the iosX64 (Intel Simulator) target --
  recent androidx.lifecycle-viewmodel and Compose Multiplatform artifacts
  no longer publish klibs for it at all, and nothing here needs it.
- Three real first-time Kotlin/Native source bugs, only surfaced once
  commonMain/iosMain were ever actually compiled for iOS: leftover
  `Clock.System.now()` calls replaced with a small `expect fun
  currentEpochMillis()` (new data/CurrentTime.kt, common/android/iOS
  actuals); an Int/Long mismatch in DateCompat.kt's lengthOfMonth(); and
  missing @OptIn(ExperimentalForeignApi::class) on IosAppFileStorage.kt's
  Foundation interop.

Full build-log-by-build-log history is in the linked Claude project's
claude/liftlog-status.md and in PORTING_PLAN.md.

Not yet exercised: actually running the app in Simulator to check the
ported UI renders and navigates correctly.

+++ shared/src/commonMain/kotlin/com/balandman/pawgress/data/DateCompat.kt
@@ -38,5 +38,12 @@ fun LocalDate.withDayOfYear(dayOfYear: Int): LocalDate = this.minusDays(this.day
 fun LocalDate.lengthOfMonth(): Int {
     val firstOfThisMonth = this.withDayOfMonth(1)
     val firstOfNextMonth = firstOfThisMonth.plusMonths(1)
-    return firstOfNextMonth.toEpochDays() - firstOfThisMonth.toEpochDays()
+    // `.toInt()` here rather than a bare subtraction: kotlinx-datetime's
+    // `toEpochDays()` returned Int as of the 0.6.1 version this project
+    // declares, but the version actually resolved for the iOS compile
+    // returned Long instead (first surfaced as a real "expected Int, actual
+    // Long" compile error in Phase 6, once Kotlin/Native ever compiled this
+    // file for the first time) -- a month is always well under Int.MAX_VALUE
+    // days, so the narrowing is safe, and it compiles under either return type.
+    return (firstOfNextMonth.toEpochDays() - firstOfThisMonth.toEpochDays()).toInt()
 }
